backend_automation/requirements.txt
firebase-admin==6.4.0
google-generativeai==0.3.2
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.3